重建测试用例完成重点使用的services，还需要补充文档和详细的测试

+++ xuper-sdk-oc/XTransaction/XTransactionOpt/XTransactionOpt.h
@@ -32,6 +32,6 @@
 @property (nonatomic, strong)   NSArray<XTranctionToAccountData*> * _Nullable to;
 @property (nonatomic, copy)     XBigInt *                           _Nullable fee;
 @property (nonatomic, strong)   XTransactionDesc *                  _Nullable desc;
-@property (nonatomic, assign)   int64_t                                       frozenHeight;
+@property (nonatomic, assign)   NSUInteger                                    frozenHeight;
 
 @end
